feat: change arrived time capsules screen ui
unify favorite & arrived time capsule state
extract time capsule list item to component
