+++ app/src/main/java/com/emotionstorage/emotionstorage/ui/AppNavHost.kt
@@ -17,9 +17,9 @@ import com.emotionstorage.auth.ui.LoginScreen
 import com.emotionstorage.domain.model.User.AuthProvider
 import com.emotionstorage.home.ui.HomeScreen
 import com.emotionstorage.my.ui.MyScreen
-import com.emotionstorage.time_capsule.ui.ArrivedScreen
+import com.emotionstorage.time_capsule.ui.ArrivedTimeCapsulesScreen
 import com.emotionstorage.time_capsule.ui.CalendarScreen
-import com.emotionstorage.time_capsule.ui.FavoritesScreen
+import com.emotionstorage.time_capsule.ui.FavoriteTimeCapsulesScreen
 import com.emotionstorage.time_capsule_detail.ui.TimeCapsuleDetailScreen
 import com.emotionstorage.tutorial.ui.OnBoardingNavHost
 import com.emotionstorage.tutorial.ui.SplashScreen
@@ -158,7 +158,7 @@ internal fun AppNavHost(
             }
 
             composable<AppDestination.ARRIVED_TIME_CAPSULES> {
-                ArrivedScreen(
+                ArrivedTimeCapsulesScreen(
                     navToTimeCapsuleDetail = { id ->
                         navController.navigate(AppDestination.TIME_CAPSULE_DETAIL(id))
                     },
@@ -168,7 +168,7 @@ internal fun AppNavHost(
             }
 
             composable<AppDestination.FAVORITE_TIME_CAPSULES> {
-                FavoritesScreen(
+                FavoriteTimeCapsulesScreen(
                     navToTimeCapsuleDetail = { id ->
                         navController.navigate(AppDestination.TIME_CAPSULE_DETAIL(id))
                     },
