+++ feat/time-capsule/src/main/java/com/emotionstorage/time_capsule/ui/FavoriteTimeCapsulesScreen.kt
@@ -2,8 +2,6 @@ package com.emotionstorage.time_capsule.ui
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.gestures.Orientation
 import androidx.compose.foundation.gestures.scrollable
 import androidx.compose.foundation.layout.Arrangement
@@ -14,30 +12,22 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.emotionstorage.common.formatToKorDateTime
 import com.emotionstorage.domain.model.TimeCapsule.Emotion
-import com.emotionstorage.time_capsule.ui.component.EmotionTag
-import com.emotionstorage.time_capsule.ui.model.FavoriteTimeCapsule
-import com.emotionstorage.ui.component.RoundedToggleButton
+import com.emotionstorage.time_capsule.ui.component.TimeCapsuleItem
+import com.emotionstorage.time_capsule.ui.model.TimeCapsuleState
 import com.emotionstorage.ui.component.TopAppBar
 import com.emotionstorage.ui.theme.MooiTheme
 import com.emotionstorage.ui.R
@@ -45,8 +35,8 @@ import com.emotionstorage.ui.component.DropDownPicker
 import java.time.LocalDateTime
 
 
-private val DUMMY_FAVORITES = (1..15).toList().map { it ->
-    FavoriteTimeCapsule(
+private val DUMMY_TIME_CAPSULES = (1..15).toList().map { it ->
+    TimeCapsuleState(
         id = it.toString(),
         title = "오늘 아침에 친구를 만났는데, 친구가 늦었어..",
         emotions = listOf(
@@ -68,13 +58,13 @@ private val DUMMY_FAVORITES = (1..15).toList().map { it ->
 }
 
 @Composable
-fun FavoritesScreen(
+fun FavoriteTimeCapsulesScreen(
     modifier: Modifier = Modifier,
     navToTimeCapsuleDetail: (id: String) -> Unit = {},
     navToBack: () -> Unit = {}
 ) {
-    StatelessFavoriteScreen(
-        favoriteTimeCapsules = DUMMY_FAVORITES,
+    StatelessFavoriteTimeCapsulesScreen(
+        timeCapsules = DUMMY_TIME_CAPSULES,
         modifier = modifier,
         navToTimeCapsuleDetail = navToTimeCapsuleDetail,
         navToBack = navToBack
@@ -82,9 +72,9 @@ fun FavoritesScreen(
 }
 
 @Composable
-private fun StatelessFavoriteScreen(
+private fun StatelessFavoriteTimeCapsulesScreen(
     modifier: Modifier = Modifier,
-    favoriteTimeCapsules: List<FavoriteTimeCapsule> = emptyList(),
+    timeCapsules: List<TimeCapsuleState> = emptyList(),
     navToTimeCapsuleDetail: (id: String) -> Unit = {},
     navToBack: () -> Unit = {}
 ) {
@@ -147,12 +137,12 @@ private fun StatelessFavoriteScreen(
                     .fillMaxWidth()
                     .scrollable(scrollState, orientation = Orientation.Vertical)
             ) {
-                items(items = favoriteTimeCapsules, key = { it.id }) {
-                    FavoriteItem(
+                items(items = timeCapsules, key = { it.id }) {
+                    TimeCapsuleItem(
                         modifier = Modifier
                             .fillMaxWidth()
                             .padding(bottom = 17.dp),
-                        item = it,
+                        timeCapsule = it,
                         onClick = { navToTimeCapsuleDetail(it.id) }
                     )
                 }
@@ -161,101 +151,13 @@ private fun StatelessFavoriteScreen(
     }
 }
 
-@Composable
-private fun FavoriteItem(
-    modifier: Modifier = Modifier,
-    item: FavoriteTimeCapsule,
-    onFavoriteClick: () -> Unit = {},
-    onClick: () -> Unit = {}
-) {
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .background(MooiTheme.colorScheme.background),
-        verticalArrangement = Arrangement.spacedBy(10.dp)
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            Text(
-                text = item.createdAt.formatToKorDateTime(),
-                style = MooiTheme.typography.body3.copy(fontWeight = FontWeight.Light),
-                color = MooiTheme.colorScheme.gray300
-            )
-            RoundedToggleButton(
-                isSelected = item.isFavorite,
-                onSelect = onFavoriteClick
-            )
-        }
-
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(92.dp)
-                .background(
-                    Color(0x1A849BEA), RoundedCornerShape(15.dp)
-                )
-                .padding(vertical = 18.dp, horizontal = 15.dp)
-                .clickable(onClick = onClick),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(11.dp)
-        ) {
-            // unlocked icon
-            Column(
-                modifier = Modifier
-                    .size(54.dp)
-                    .border(1.dp, Color(0xAECBFA).copy(alpha = 0.2f), CircleShape),
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Image(
-                    painter = painterResource(id = R.drawable.lock_open),
-                    modifier = Modifier
-                        .width(11.dp)
-                        .height(14.dp),
-                    contentDescription = "open"
-                )
-                Text(
-                    modifier = Modifier.padding(top = 3.dp),
-                    text = "열림",
-                    style = MooiTheme.typography.body3.copy(fontSize = 11.sp, lineHeight = 24.sp),
-                    color = MooiTheme.colorScheme.secondary
-                )
-            }
-
-            // time capsule content
-            Column(
-                modifier = Modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(5.dp)
-            ) {
-                Row(
-                    horizontalArrangement = Arrangement.spacedBy(4.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    for (emotion in item.emotions) {
-                        EmotionTag(emotion = emotion)
-                    }
-                }
-                Text(
-                    text = item.title,
-                    style = MooiTheme.typography.body4,
-                    color = MooiTheme.colorScheme.primary,
-                    maxLines = 1
-                )
-            }
-        }
-    }
-}
-
 
 @Preview
 @Composable
-private fun FavoriteScreenPreview() {
+private fun FavoriteTimeCapsulesScreenPreview() {
     MooiTheme {
-        StatelessFavoriteScreen(
-            favoriteTimeCapsules = DUMMY_FAVORITES
+        StatelessFavoriteTimeCapsulesScreen(
+            timeCapsules = DUMMY_TIME_CAPSULES
         )
     }
 }
