--- feat/time-capsule/src/main/java/com/emotionstorage/time_capsule/ui/model/ArrivedTimeCapsule.kt
@@ -1,11 +1,0 @@
-package com.emotionstorage.time_capsule.ui.model
-
-import com.emotionstorage.domain.model.TimeCapsule.Emotion
-import java.time.LocalDateTime
-
-data class ArrivedTimeCapsule(
-    val id: String,
-    val title: String,
-    val emotions: List<Emotion> = emptyList(),
-    val arriveAt: LocalDateTime,
-)
