+++ feat/time-capsule/src/main/java/com/emotionstorage/time_capsule/ui/model/TimeCapsuleState.kt
@@ -3,7 +3,7 @@ package com.emotionstorage.time_capsule.ui.model
 import com.emotionstorage.domain.model.TimeCapsule.Emotion
 import java.time.LocalDateTime
 
-data class FavoriteTimeCapsule (
+data class TimeCapsuleState (
     val id: String,
     val title: String,
     val emotions: List<Emotion> = emptyList(),
